+++ users/src/lib/config.ts
@@ -55,7 +55,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = Object.freeze({
     language: "typescript",
     openapiDocVersion: "1.0.0",
-    sdkVersion: "2.0.1",
-    genVersion: "2.275.2",
-    userAgent: "speakeasy-sdk/typescript 2.0.1 2.275.2 1.0.0 @wingspan/users",
+    sdkVersion: "2.0.2",
+    genVersion: "2.277.0",
+    userAgent: "speakeasy-sdk/typescript 2.0.2 2.277.0 1.0.0 @wingspan/users",
 });